Add files via upload
Despite the fact that it was my third attempt, I managed to score 0.45 in my second submission on this kaggle competition. A lot better than my first one.

This time a took some time to start from scratch and make an objective view of the most relevant characteristics that affected the price of the houses, despite from neighborhood, which in my opinion is the most important, but I wasn't able to qualify it yet. Maybe for a fourth try, which is very unlikely since I scored well.